+++ mnp/licenses/MIT.js
@@ -1,6 +1,0 @@
-/*!
- * my-new-package
- * {{ repo.html_url }}
- * Copyright(c) {{ year }} {{ legal_name }}
- * MIT Licensed
- */