+++ types/externs.js
@@ -5,26 +5,26 @@
 
 /* typal types/index.xml externs */
 /** @const */
-var _myNewPackage = {}
+var _hotReloadArticle = {}
 /**
  * Options for the program.
  * @record
  */
-_myNewPackage.Config
+_hotReloadArticle.Config
 /**
  * A boolean option. Default `true`.
  * @type {boolean|undefined}
  */
-_myNewPackage.Config.prototype.shouldRun
+_hotReloadArticle.Config.prototype.shouldRun
 /**
  * A text to return.
  * @type {string|undefined}
  */
-_myNewPackage.Config.prototype.text
+_hotReloadArticle.Config.prototype.text
 
 /* typal types/api.xml externs */
 /**
- * {{ description }}
- * @typedef {function(!_myNewPackage.Config): !Promise<string>}
+ * An article about hot reload.
+ * @typedef {function(!_hotReloadArticle.Config): !Promise<string>}
  */
-_myNewPackage.myNewPackage
+_hotReloadArticle.hotReloadArticle