+++ types/index.js
@@ -2,10 +2,10 @@ export {}
 
 /* typal types/api.xml namespace */
 /**
- * @typedef {_myNewPackage.myNewPackage} myNewPackage {{ description }}
- * @typedef {(config: !_myNewPackage.Config) => string} _myNewPackage.myNewPackage {{ description }}
+ * @typedef {_hotReloadArticle.hotReloadArticle} hotReloadArticle An article about hot reload.
+ * @typedef {(config: !_hotReloadArticle.Config) => string} _hotReloadArticle.hotReloadArticle An article about hot reload.
  */
 
 /**
- * @typedef {import('..').Config} _myNewPackage.Config
+ * @typedef {import('..').Config} _hotReloadArticle.Config
  */