+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-## {{ create_date }}
+## 4 March 2020
 
 ### 0.0.0-pre
 
-- Create `{{ package-name }}` with _[`NodeTools`](https://art-deco.github.io/nodetools)_.
+- Create `@idio/hot-reload-article` with _[`NodeTools`](https://art-deco.github.io/nodetools)_.